Split plwf105 firmware into individual files
I have a number of esphome devices and so I have a more advanced firmware setup using a common file defining things like time sources and network access. The current project file is a complete esphome firmware definition and so it's a little hard to use it as it currently stands. I have to import it in and then !remove a bunch of values and sensors I don't want.

I went and split up the yaml file into functional chunks and features to allow easier advanced firmware. One could import individual functional chunks or features, or still import the main project yaml and get the entire thing.

I've been running the functional imports for the last day or so and it appears that I didn't break anything, but it's a large change and so I'm sure I might have messed up something in the split and 'cleanup'.

Appreciate your consideration for the changes and thank you very much for the overall project!